pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "transforna"
authors = [
    {name = "Yasser Taha", email = "ytaha@hb-dx.com"},
    {name = "Julia Jehn", email = "jjehn@hb-dx.com"},
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]

dependencies = [
"anndata==0.8.0",
"dill==0.3.6",
"hydra-core==1.3.0",
"imbalanced-learn==0.9.1",
"matplotlib==3.5.3",
"numpy==1.22.3",
"omegaconf==2.2.2",
"pandas==1.5.2",
"plotly==5.10.0",
"PyYAML==6.0",
"rich==12.6.0",
"viennarna>=2.5.0a5",
"scanpy==1.9.1",
"scikit_learn==1.2.0",
"skorch==0.12.1",
"torch==1.10.1",
"tensorboard==2.11.2",
"Levenshtein==0.21.0"]